GetVelEps

FRGn.GetVelEps.get_velocity — Function

get_velocity(velocity::Array{Float64,2}, k1::Float64, k2::Float64, cutoff::Float64, m::Int64, n::Int64)

Returns the velocities at momenta k1 and k2 at the previous cutoff (indexed m-i+2).

get_velocity(velocity::Array{Float64,2}, k::Float64, cutoff::Float64, m::Int64, n::Int64)

Returns the velocities at momenta k at the previous cutoff (indexed m-i+2).

FRGn.GetVelEps.get_dielectric — Function

get_dielectric(dielctric::Array{Float64,2}, k1::Float64, k2::Float64, cutoff::Float64, m::Int64, n::Int64)

Returns the dielectric functions at momenta k1 and k2 at the previous cutoff (indexed m-i+2).

get_dielectric(dielctric::Array{Float64,2}, k::Float64,cutoff::Float64, m::Int64, n::Int64)

Returns the dielectric functions at momenta k at the previous cutoff (indexed m-i+2).

Missing docstring.

Missing docstring for fetch_value. Check Documenter's build log for details.