Sample Functions

Sample Functions

Bauer

BosonInShell

Finite Temperature

Bauer

FRGn.Bauer.velocity_integrand — Function

velocity_integrand(velocity::Array{Float64,2},dielectric::Array{Float64,2}, momentum::Float64, cutoff::Float64, phi::Float64, m::Int64, n::Int64, i::Int64)

This function returns the integrand of the FRG equation for the velocity renormlisation.

Args

dielectric (Array) : Array containing the dielectric values
momentum (Float64) : momentum value
cutoff   (FLoat64) : running cutoff
phi      (Float64) : angular coordinate
m          (Int64) : number of cutoffs
n          (Int64) : number of momenta
i          (Int64) : index for the running cutoff

FRGn.Bauer.dielectric_integrand — Function

dielectric_integrand(velocity::Array{Float64,2},dielectric::Array{Float64,2}, momentum::Float64, cutoff::Float64, phi::Float64, m::Int64, n::Int64, i::Int64)

This function returns the integrand of the FRG equation for the dielectric function renormlisation.

Args

velocity   (Array) : The velocity array
momentum (Float64) : momentum value
cutoff   (FLoat64) : running cutoff
phi      (Float64) : angular coordinate
m          (Int64) : number of cutoffs
n          (Int64) : number of momenta
i          (Int64) : index for the running cutoff

BosonInShell

FRGn.BosonInShell.velocity_integrand — Function

velocity_integrand(velocity::Array{Float64,2},dielectric::Array{Float64,2}, momentum::Float64, cutoff::Float64, phi::Float64, m::Int64, n::Int64)

This function returns the integrand of the FRG equation for the velocity renormlisation.

Args

velocity   (Array) : The velocity array
dielectric (Array) : Array containing the dielectric values
momentum (Float64) : momentum value
cutoff   (FLoat64) : running cutoff
phi      (Float64) : angular coordinate (should run from 0 to pi/2)
m          (Int64) : number of cutoffs
n          (Int64) : number of momenta

FRGn.BosonInShell.dielectric_integrand — Function

dielectric_integrand(velocity::Array{Float64,2},dielectric::Array{Float64,2}, momentum::Float64, cutoff::Float64, phi::Float64, m::Int64, n::Int64)

This function returns the integrand of the FRG equation for the dielectric function renormlisation.

Args

velocity   (Array) : The velocity array
dielectric (Array) : Array containing the dielectric values
momentum (Float64) : momentum value
cutoff   (FLoat64) : running cutoff
phi      (Float64) : angular coordinate (should run from 0 to pi/2)
m          (Int64) : number of cutoffs
n          (Int64) : number of momenta

Finite Temperature

FRGn.FiniteTemp.velocity_integrand_t — Function

velocity_integrand_t(velocity::Array{Float64,2},dielectric::Array{Float64,2}, momentum::Float64, cutoff::Float64, phi::Float64, m::Int64, n::Int64, temp::Float64)

This function returns the integrand of the FRG equation for the velocity renormlisation at a specific temperature.

Args

velocity   (Array) : The velocity array
dielectric (Array) : Array containing the dielectric values
momentum (Float64) : momentum value
cutoff   (FLoat64) : running cutoff
phi      (Float64) : angular coordinate (should run from 0 to pi/2)
m          (Int64) : number of cutoffs
n          (Int64) : number of momenta
i          (Int64) : index for the running cutoff
temp     (Float64) : temperature

FRGn.FiniteTemp.dielectric_integrand_t — Function

dielectric_integrand_t(velocity::Array{Float64,2},dielectric::Array{Float64,2}, momentum::Float64, cutoff::Float64, phi::Float64, m::Int64, n::Int64, temp::Float64)

This function returns the integrand of the FRG equation for the dielectric function renormlisation at a specific temperature.

Args

velocity   (Array) : The velocity array
dielectric (Array) : Array containing the dielectric values
momentum (Float64) : momentum value
cutoff   (FLoat64) : running cutoff
phi      (Float64) : angular coordinate
m          (Int64) : number of cutoffs
n          (Int64) : number of momenta
i          (Int64) : index for the running cutoff
temp     (Float64) : temperature